requirements.txt
langchain
pinecone-client
streamlit
openai
python-dotenv
tiktoken
pypdf
guidance